cli mod reorg
